Moves UpdateMarcRecordService and cleans up rubocop violations.
